Updating Routes

+++ celestial-charm-csp/server/routes/users.js
@@ -3,11 +3,33 @@ const router = Router();
 import User from '../models/User.js';
 import psd from 'bcryptjs';
 const { genSalt, hash } = psd;
-import  { requireAuth }  from '../middleware/requireAuth.js';
+import { requireAuth } from '../middleware/requireAuth.js';
 
 const JWT_SECRET = process.env.JWT_SECRET || 'AteezPresent';
 
+if (!JWT_SECRET) {
+    throw new Error(
+        "JWT_SECRET is required"
+    );
+}
+
+router.get('/me', requireAuth, async (req, res) => {
+    try {
+        const user = await User.findById(req.user.id)
+        // .select("name username email phoneNumber birthday profilePicture gems personalityType inventory");
+        if (!user) return res.status(404).json({ error: 'User not found' });
+        res.json(user);
+    } catch (err) {
+        res.status(500).json({ error: 'Server error' });
+    }
+});
+
 router.get('/:id', requireAuth, async (req, res) => {
+    if (req.params.id !== req.user.id) {
+        return res.status(403).json({
+            message: "Forbidden"
+        });
+    }
     try {
         const user = await User.findById(req.params.id);
         if (!user) return res.status(404).json({ error: 'User not found' });
@@ -19,6 +41,11 @@ router.get('/:id', requireAuth, async (req, res) => {
 });
 
 router.put('/:id', requireAuth, async (req, res) => {
+    if (req.params.id !== req.user.id) {
+        return res.status(403).json({
+            message: "Forbidden"
+        });
+    }
     try {
         const updates = { ...req.body };
 
@@ -39,6 +66,11 @@ router.put('/:id', requireAuth, async (req, res) => {
 });
 
 router.delete('/:id', requireAuth, async (req, res) => {
+    if (req.params.id !== req.user.id) {
+        return res.status(403).json({
+            message: "Forbidden"
+        });
+    }
     try {
         await User.findByIdAndDelete(req.params.id);
         res.status(200).json({ message: 'Account deleted' });
@@ -47,19 +79,5 @@ router.delete('/:id', requireAuth, async (req, res) => {
         res.status(500).json({ error: 'Failed to delete account.' });
     }
 });
-
-router.get('/me', requireAuth, async (req, res) => {
-    try {
-        const user = await User.findById(req.user.id)
-        // .select("name username email phoneNumber birthday profilePicture gems personalityType inventory");
-        if (!user) return res.status(404).json({ error: 'User not found' });
-        res.json(user);
-    } catch (err) {
-        res.status(500).json({ error: 'Server error' });
-    }
-});
-
-
-
 
 export default router;
